by default, define and use the unaligned versions of the eigen types
The "normal" eigen types cannot be used as-is in the orocos dataflow
since they have strong alignment requirements. However, it seems that
for small sizes (below 4), having aligned types does not give
any real benefit.

So, move to using unaligned versions of the eigen types so that we
can integrate the datatypes in orocos directly without convertions.