cmake: fix handling of public dependencies
The default of adding all dependencies as public dependencies was
broken, as it was using the wrong variable names. This was hidden as
the pkgconfig handling was using the non-public variable names to
setup the requires, and not only the public ones. Since Rock is mainly
using pkg-config, the bugs did not trigger.

Fix documentation of rock_add_public_dependencies