logging: change log level in case Release or NDEBUG is defined
Before this commit, if either Release or NDEBUG are defined, any logging
commands would not be compiled in. This includes error messages, which
is generally not what you would expect. Also, the level could not be
changed by explicitly setting the compiled in debug level.

The behavior is now that if no specific level for debugging is specified
at compile time using the BASE_LOG_XXX symbols, it works the following:
If Release or NDEBUG are defined,
set to BASE_LOG_WARN Otherwise set to BASE_LOG_DEBUG (all levels)