don't shadow local variable in "Time::fromString()"
cutting away the ":millisecond" part of the string is not needed, as
"strptime()" will simply not consume characters after a successfull
match. to make matters worse, the lifetime of the newly created,
shadowing variable was only limited to be inside the if-clause and had
no effect to the rest of the function anyways...

fixes compiler warning.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>