viz: simplify sonar beam visualization code
Instead of doing its own management of the sonar stream, it now uses the
keepOldData features to achieve the same result as the existing code, but
a lot simpler (and easy to change at runtime). The beams are now
accurately represented (i.e. are plus/minus a half-width around the
center bearing). Finally, bearing=0 means beam along X.