test: rewrite the AngleSegment test
The last test (loop in a loop) was very obscure specification-wise,
and was actually broken - only that its broken-ness was not showing
up because the test writer used assert() instead of test assertions,
which were elided by the compiler since we compile with optimizations.

Rewrite the test trying to follow all the possible corner-cases.