remove dependency of the typelib types on the base typekit
This dependency is an issue because of the circular dependencies it
entails. In addition, moving these files to base/orogen/types is also
problematic since the typelib conversions can be used without the
typekit (e.g. analysing pocolog logs).

Copy the current state of `base.tlb` and use it directly. Since types
aren't meant to change often, I think it provides a good middle ground.