feat: RigidBodyStateVisualization: allow displaying inverse transform
This allows display of the coordinate system of the targetframe as viewed
from the sourceframe. Previously, the only useful way to setup the view in
regards to the other transformations was to set the frame of the
visualisation to the targetFrame, and the model/coordinate cross would
appear at the origin of the sourceFrame. Setting displayTargetInSource to
true, one would then set the visualisation frame to the sourceFrame and
get the coordinate cross(and model) at the origin of the targetFrame.